reduce file separation
This commit addresses https://github.com/reach-sh/reach-lang/pull/655#pullrequestreview-901936011.
