--- vsce/client/src/constants/hideNotificationText.ts
@@ -1,1 +1,0 @@
-export default 'Hide Reach update messages';
