--- vsce/client/src/util/checkDocker.ts
@@ -1,8 +1,0 @@
-import { spawnSync } from 'child_process';
-
-/**
- * @returns {number} `1` if Docker is **not** running.
- */
-export default (): number => spawnSync(
-	'docker', [ '--version' ]
-).status;
